240517_postgresql

+++ TodoApp/routers/todos.py
@@ -95,15 +95,23 @@ async def create_todo(user : user_dependency, db : db_dependency, todo_request :
     db.commit() # transaction
 
 @router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db : db_dependency,
+async def update_todo(user : user_dependency,
+                      db : db_dependency,
                       todo_request : TodoRequest,
                       todo_id : int = Path(gt=0),
                       #todo_request : TodoRequest # 매개변수보다 앞에 있어야함
                       ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail='Authentication Failed')
 
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).\
+        filter(Todos.owner_id == user.get('id')).first()
+    
     if todo_model is None:
-        raise HTTPException(status_code=404,detail='Todo not found')
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail='Todo not found')
     
     todo_model.title = todo_request.title
     todo_model.description = todo_request.description
@@ -114,10 +122,17 @@ async def update_todo(db : db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db : db_dependency, todo_id : int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user : user_dependency,
+                      db : db_dependency, todo_id : int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).\
+        filter(Todos.owner_id == user.get('id')).first()
+    
     if todo_model is None:
         raise HTTPException(status_code=404,detail='Todo not found.')
+    
     db.query(Todos).filter(Todos.id==todo_id).delete()
     db.commit()
